+++ editorextensions/codebeamer-custom-ui/modal/src/pages/import/components/importer/MiroImporter.tsx
@@ -3,11 +3,7 @@ import Modal from 'react-bootstrap/Modal';
 import ProgressBar from 'react-bootstrap/ProgressBar';
 import Spinner from 'react-bootstrap/Spinner';
 import { useGetItemsQuery } from '../../../../api/codeBeamerApi';
-import {
-	closeModal,
-	createAppCard,
-	startImport,
-} from '../../../../api/lucidGateway';
+import { LucidGateway } from '../../../../api/lucidGateway';
 import {
 	DEFAULT_RESULT_PAGE,
 	MAX_ITEMS_PER_IMPORT,
@@ -37,7 +33,7 @@ export default function Importer(props: {
 	React.useEffect(() => {
 		const importItems = async (items: CodeBeamerItem[]) => {
 			const importId = Math.ceil(Math.random() * 899) + 100;
-			startImport(importId, items.length);
+			LucidGateway.startImport(importId, items.length);
 			const _items: CodeBeamerItem[] = structuredClone(items);
 			for (let i = 0; i < _items.length; i++) {
 				console.log('Item no. ' + i + ' of ' + _items.length);
@@ -50,7 +46,7 @@ export default function Importer(props: {
 						continue;
 					}
 				}
-				await createAppCard(
+				await LucidGateway.createAppCard(
 					importId,
 					_items[i],
 					props.items.find(